Change default name of php-fpm script to php5-fpm
This is the new default to match AAI's ansible scripts that utilize
Ubuntu and nginx.

You can still override the default by setting the `:phpfpm_init_command` like so:

```ruby
set :phpfpm_init_command, "/etc/init.d/php-fpm"

set :phpfpm_init_command, "service php-fpm"
```

In the future, we'll try to make this a little more intelligent and sniff for which
PHP FPM package is installed and what distribution of Linux is installed.